requirements.txt
fastapi
uvicorn
celery
redis
torch
torchvision
pillow
pynvml
